messages_service/requirements.txt
Flask==2.0.3
pika==1.2.1
python-consul2==0.1.5